GHTRequester: Make parsing of response stateful
This replaces get:as: with responseAs:, which does the same as what get:as: did before, except that it now works in a stateful way, using the previous response returned by ZnClient. This way we can parse responses of not just GET requests, but also PATCH, DELETE, etc. This commit includes also the refactoring to use responseAs: instead.