Add method performing responseAs: with a temporary error handler
This adds responseAs:onInvalidFields: which temporarily sets the error handler for invalid fields, as this is an error which occurs for many parts of the API for different reasons. This allows other parts of the API to throw a more special kind of error (e.g. 'Pull Request already exists'), or handle the error by continuing with a different strategy.