Add more tests for repeated invocations of #isOutdated.
Sending #update didn't work if #isOutdated was sent beforehand.
This is because #isOutdated modifies the ETag.

The solution is to keep let each object keep track of its own ETag,
and then comparing that.

For this the tests had to be adapted so that each response contains
a URL (to propagate the ETag, not committed as of this commit).